Created staticwebapp config & componentized the portion we want hidden
Put the portion of the code we are going to have behind the login mechanism into its own portion to make index.astro easier to parse. Implemented the Azure configuration, at the moment just to redirect the user from the app if they are not "authorized" (logged in) Going to try what Max suggested and create a PR to test to see if the login mechanism works this simply.