Added prettier and implemented a login status on generate page
Implemented prettier plugin to correct a lot of our syntax structure. That way things are properly aligned how they should be, and automatically works with the plugin if you have it to fix on ctrl + s. I also implemented a login specification on the generate page so you can tell if you are logged in. (David requested this a few meetings back, even though it is not on the XLS file), and its a good way to test the dev branch working as expected.